// nicnbk-ui-angular2/src/app/m2s2/model/re-memo.ts
import {FundMemo} from "./fund-memo";

export class REMemo extends FundMemo{

    conviction: number;
    
    teamNotes: string;
    teamScore: number;
    trackRecordNotes: string;
    trackRecordScore: number;
    strategyNotes: string;
    strategyScore: number;
    otherInfo: string;
    NICFollowups: string;
    otherFollowups: string;

}